@@ -1,6 +1,6 @@
 package carcrowdsystem.ccs.entitys;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.util.List;
 
 @Entity
